location details component added
