--- src/components/index.js
@@ -1,7 +1,0 @@
-export { default as Header } from "./Header";
-export { default as LoginForm } from "./LoginForm";
-export { default as SignUpForm } from "./SignUpForm";
-export { default as Footer } from "./Footer";
-export { default as LandingPage } from "./LandingPage";
-export { default as Paths } from "./Paths";
-export { default as UserProfile } from "./UserProfile";
