+++ src/index.js
@@ -2,28 +2,11 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import "./styles/index.css";
-import {
-  Header,
-  Footer,
-  LoginForm,
-  SignUpForm,
-  LandingPage,
-  UserProfile,
-  Paths,
-} from "./components";
+import App from "../src/App";
 
 ReactDOM.render(
-  <Router>
-    <Header />
-    <Routes>
-      <Route path="/" element={<LandingPage />} />
-      <Route path="/signUpForm" element={<SignUpForm />} />
-      <Route path="/loginForm" element={<LoginForm />} />
-      <Route path="/userProfile" element={<UserProfile />} />
-      <Route path="/paths" element={<Paths />} />
-    </Routes>
-    <Footer />
-  </Router>,
-
-  document.getElementById("root")
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
 );
